refactor: bake light map (#162)
