+++ packages/custom-material/src/bake-pbr/vertex.ts
@@ -1,0 +1,28 @@
+export default `
+#include <common>
+#include <common_vert>
+#include <blendShape_input>
+#include <uv_share>
+#include <color_share>
+#include <normal_share>
+#include <worldpos_share>
+
+#include <ShadowVertexDeclaration>
+#include <FogVertexDeclaration>
+
+void main() {
+
+    #include <begin_position_vert>
+    #include <begin_normal_vert>
+    #include <blendShape_vert>
+    #include <skinning_vert>
+    #include <uv_vert>
+    #include <color_vert>
+    #include <normal_vert>
+    #include <worldpos_vert>
+    #include <position_vert>
+
+    #include <ShadowVertex>
+    #include <FogVertex>
+}
+`;
